Decrease number of threads for linkchecker to 21
Default is 10 and I think it is what could contribute to linkchecker failing to
even establish connection to those websites. We tried "2" and it still failed
on 1 url, thus reducing to 1.

Also relevant issue/discussion is absence in difference of unable to open connection
to actually finding no webpage there:
https://github.com/linkchecker/linkchecker/issues/169